Rename resources to apis
